Improvement of the function of getting the file size from the status bar.
This function is necessary when it is impossible to obtain the file size by other means, for example, when working via FTP or MTP protocol. Now there should be fewer errors, and the file size will be transferred as displayed by Total Commander. The size will be converted to bytes if no units are specified, otherwise no action is required. In TC settings, you can set any size display format that will be taken into account by this function.